+++ src/wallet-api/typedData.ts
@@ -1,4 +1,23 @@
+import { ADDRESS, FELT } from '../api/components.js';
 import type { ABI_TYPE_ENUM } from '../api/constants.js';
+
+/**
+ * A string representing an unsigned integer
+ * @pattern ^(0|[1-9]{1}[0-9]*)$
+ */
+export type NUMERIC = string;
+/**
+ * UNIX time
+ */
+export type TIMESTAMP = NUMERIC;
+
+/**
+ * Object containing timestamps corresponding to `Execute After` and `Execute Before`
+ */
+export type TIME_BOUNDS = {
+  execute_after: TIMESTAMP;
+  execute_before: TIMESTAMP;
+};
 
 export const TypedDataRevision = {
   ACTIVE: '1',
@@ -54,3 +73,46 @@ export interface TypedData {
   domain: StarknetDomain;
   message: object;
 }
+
+/**
+ * "A typed data object (in the sense of SNIP-12) which represents an outside execution payload, according to SNIP-9
+ */
+export type OutsideExecutionTypedData = OutsideExecutionTypedDataV1 | OutsideExecutionTypedDataV2;
+export type OutsideExecutionTypedDataV1 = {
+  types: Record<string, StarknetType[]>;
+  primaryType: string;
+  domain: StarknetDomain;
+  message: OutsideExecutionMessageV1;
+};
+export type OutsideExecutionTypedDataV2 = {
+  types: Record<string, StarknetType[]>;
+  primaryType: string;
+  domain: StarknetDomain;
+  message: OutsideExecutionMessageV2;
+};
+export type OutsideExecutionMessageV1 = {
+  caller: FELT;
+  nonce: FELT;
+  execute_after: FELT;
+  execute_before: FELT;
+  calls_len: FELT;
+  calls: OutsideCallV1[];
+};
+export type OutsideCallV1 = {
+  to: ADDRESS;
+  selector: FELT;
+  calldata_len: FELT[];
+  calldata: FELT[];
+};
+export type OutsideExecutionMessageV2 = {
+  Caller: FELT;
+  Nonce: FELT;
+  'Execute After': FELT;
+  'Execute Before': FELT;
+  Calls: OutsideCallV2[];
+};
+export type OutsideCallV2 = {
+  To: ADDRESS;
+  Selector: FELT;
+  Calldata: FELT[];
+};
